+++ src/utils/types.ts
@@ -1,0 +1,11 @@
+export type SiteProps = {
+  url: string;
+  title: string;
+  author: string;
+  description?: string;
+  defaultImageUrl?: string;
+  wpGraphQlUrl?: string;
+}
+
+
+